# optimize alloc