--- a/random/random_string.go
+++ b/random/random_string.go
@@ -1,5 +1,7 @@
 package random
 
+var defaultRandString = newRandString()
+
 type randString struct {
 	randBytes    []byte
 	randBytesLen int