goos: darwin
goarch: arm64
pkg: go.x2ox.com/sorbifolia/random
BenchmarkFastRand
BenchmarkFastRand-10           	  337707	      3526 ns/op
BenchmarkMathRand
BenchmarkMathRand-10           	  237457	      5219 ns/op
BenchmarkFastRand_Number
BenchmarkFastRand_Number-10    	17673829	        66.48 ns/op
BenchmarkSafeRand_Number
BenchmarkSafeRand_Number-10    	  110781	     10749 ns/op
BenchmarkMathRand_Number
BenchmarkMathRand_Number-10    	 8781788	       137.2 ns/op
BenchmarkSafeRand
BenchmarkSafeRand-10           	  256386	      4518 ns/op
PASS